chg: Initial swap of Hyperscan libraries with Vectorscan.
