+++ .github/pull_request_template.md
@@ -5,7 +5,7 @@ Please confirm the following:
 - [ ] I have run tests locally, and they all pass.
 - [ ] I have added or extended tests, to cover any new features or changes included in this PR.
 - [ ] I have added or updated documentation, to cover any new features or changes included in this PR.
-- [ ] I have followed the [Contributing Guide](/pyranha-labs/hypergrep/blob/main/CONTRIBUTING.md) to ensure code quality to the best of my ability.
+- [ ] I have followed the [Contributing Guide](/pyranha-labs/vectorgrep/blob/main/CONTRIBUTING.md) to ensure code quality to the best of my ability.
 - [ ] I have self-reviewed my PR to ensure the guidelines have been followed to the best of my ability.
 
 If you need assistance before the checklist can be completed, please add a comment with detailed
