+++ MANIFEST.in
@@ -1,8 +1,9 @@
 include README.md
 include LICENSE
+include COPYING
 include requirements*.txt
-include hypergrep/lib/libhs.so.5.4.2
-include hypergrep/lib/libhyperscanner.so
-include hypergrep/lib/libzstd.so.1.5.5
-include hypergrep/lib/c/hyperscanner.c
-include hypergrep/lib/c/hyperscanner.h
+include vectorgrep/lib/libhs.so.5.4.11
+include vectorgrep/lib/libvectorgrep.so
+include vectorgrep/lib/libzstd.so.1.5.5
+include vectorgrep/lib/c/vectorgrep.c
+include vectorgrep/lib/c/vectorgrep.h
