+++ bandit.yaml
@@ -7,4 +7,4 @@ exclude_dirs:
     - tests
 
 targets:
-  - hypergrep
+  - vectorgrep
