--- hypergrep/lib/__init__.py
@@ -1,1 +1,0 @@
-"""Shared files and libraries across languages for reading raw text files with Intel Hyperscan."""
