--- hypergrep/test/samplefile.txt
@@ -1,4 +1,0 @@
-foo
-foobar
-barfoo
-food
