+++ setup.cfg
@@ -24,7 +24,7 @@ disable=raw-checker-failed,
         useless-suppression,
         deprecated-pragma,
         use-symbolic-message-instead,
-        # HyperGrep specific modifications.
+        # VectorGrep specific modifications.
         # W1203, allow f-strings to be used in logging messages.
         logging-fstring-interpolation,
         # C0301 line too long, up to maintainers/contributors to use best judgement.
