+++ vectorgrep/__init__.py
@@ -1,0 +1,15 @@
+"""Libraries for reading raw text files with Vectorscan."""
+
+from vectorgrep.utils import CALLBACK_TYPE
+from vectorgrep.utils import HS_FLAG_CASELESS
+from vectorgrep.utils import HS_FLAG_DOTALL
+from vectorgrep.utils import HS_FLAG_MULTILINE
+from vectorgrep.utils import HS_FLAG_SINGLEMATCH
+from vectorgrep.utils import Result
+from vectorgrep.utils import check_compatibility
+from vectorgrep.utils import configure_libraries
+from vectorgrep.utils import grep
+from vectorgrep.utils import prepare_patterns
+from vectorgrep.utils import scan
+
+__version__ = "0.0.1"
