+++ vectorgrep/lib/__init__.py
@@ -1,0 +1,1 @@
+"""Shared files and libraries across languages for reading raw text files with Vectorscan."""
