+++ vectorgrep/test/samplefile.txt
@@ -1,0 +1,4 @@
+foo
+foobar
+barfoo
+food
